@@ -1,10 +1,5 @@
-import React from 'react';
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Link
-} from "react-router-dom";
+import React from "react";
+import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
 
 import { StateProvider } from "./hooks/useGlobalState";
 import { reducer } from "./utilities/reducer";
@@ -14,15 +9,14 @@ import Settings from "./components/settings";
 import Storage from "./utilities/storage";
 
 const App = () => {
-
   const storage = new Storage();
 
   const getTeams = () => {
     //check storage
-    if( storage.retrieve(`standupTimer`) ){
+    if (storage.retrieve(`standupTimer`)) {
       let store = storage.retrieve(`standupTimer`);
-      if( store.items.teams ){
-        if( store.items.teams.length > 0 ){
+      if (store.items.teams) {
+        if (store.items.teams.length > 0) {
           return store.items.teams;
         }
       }
@@ -30,80 +24,99 @@ const App = () => {
 
     return [
       {
-        name: 'My Team',
+        name: "My Team",
         members: [
           {
-            name: 'Team Member Here...',
+            name: "Team Member Here...",
             active: true,
-            speech: 'Team Member',
-            mediaUrl: null
-          }
-        ]
+            speech: "Team Member",
+            mediaUrl: null,
+          },
+        ],
       },
       {
-        name: 'My Second Team',
+        name: "My Second Team",
         members: [
           {
-            name: 'Team Member Here...',
+            name: "Team Member Here...",
             active: true,
-            speech: 'Team Member',
-            mediaUrl: null
-          }
-        ]
-      }
-    ]
-  }
+            speech: "Team Member",
+            mediaUrl: null,
+          },
+        ],
+      },
+    ];
+  };
 
   const getSettings = (index) => {
-    if( storage.retrieve(`standupTimer`) ){
+    if (storage.retrieve(`standupTimer`)) {
       let store = storage.retrieve(`standupTimer`);
-      if( store.items.settings ){
+      if (store.items.settings) {
         return store.items.settings;
       }
     }
 
     return {
       origDuration: 60000,
-      gongMessage: ['Please stop'],
+      gongMessage: ["Please stop"],
       gongMediaUrl: [],
     };
-  }
+  };
 
-  let colors = ( typeof window !== 'undefined' && typeof document !== 'undefined' ) ? {
-    white: '#fff',
-    backgroundColor: window.getComputedStyle(document.documentElement).getPropertyValue('--background-color'),
-    highlightOne: window.getComputedStyle(document.documentElement).getPropertyValue('--highlight-one'),
-    lowlightOne: window.getComputedStyle(document.documentElement).getPropertyValue('--lowlight-one'),
-    vanilla: window.getComputedStyle(document.documentElement).getPropertyValue('--vanilla'),
-    flax: window.getComputedStyle(document.documentElement).getPropertyValue('--flax'),
-    neonCarrot: window.getComputedStyle(document.documentElement).getPropertyValue('--neon-carrot'),
-    giantsOrange: window.getComputedStyle(document.documentElement).getPropertyValue('--giants-orange'),
-    watermelonRed: window.getComputedStyle(document.documentElement).getPropertyValue('--watermelon-red'),
-  } : {} ;
+  let colors =
+    typeof window !== "undefined" && typeof document !== "undefined"
+      ? {
+          white: "#fff",
+          backgroundColor: window
+            .getComputedStyle(document.documentElement)
+            .getPropertyValue("--background-color"),
+          highlightOne: window
+            .getComputedStyle(document.documentElement)
+            .getPropertyValue("--highlight-one"),
+          lowlightOne: window
+            .getComputedStyle(document.documentElement)
+            .getPropertyValue("--lowlight-one"),
+          vanilla: window
+            .getComputedStyle(document.documentElement)
+            .getPropertyValue("--vanilla"),
+          flax: window
+            .getComputedStyle(document.documentElement)
+            .getPropertyValue("--flax"),
+          neonCarrot: window
+            .getComputedStyle(document.documentElement)
+            .getPropertyValue("--neon-carrot"),
+          giantsOrange: window
+            .getComputedStyle(document.documentElement)
+            .getPropertyValue("--giants-orange"),
+          watermelonRed: window
+            .getComputedStyle(document.documentElement)
+            .getPropertyValue("--watermelon-red"),
+        }
+      : {};
 
   const initialState = {
-    mode: 'start',
+    mode: "start",
     teams: getTeams(),
     currentTeam: 0,
     colors: colors,
-    imageExtensions: ['gif','jpg','jpeg','png'],
-    videoExtensions: ['mp4','mov'],
+    imageExtensions: ["gif", "jpg", "jpeg", "png"],
+    videoExtensions: ["mp4", "mov"],
     settings: getSettings(),
-  }
+  };
 
   console.log(getTeams());
 
   return (
     <StateProvider reducer={reducer} initialState={initialState}>
       <Router>
-        <Switch>
-          <Route exact path="/" component={StandupTimer} />
-          <Route path="/settings" component={Settings} />
-          <Route component={ () => (<div>404 Not Found!</div>)} />
-        </Switch>
+        <Routes>
+          <Route exact path="/" element={StandupTimer()} />
+          <Route path="/settings" element={Settings} />
+          <Route element={() => <div>404 Not Found!</div>} />
+        </Routes>
       </Router>
     </StateProvider>
   );
-}
+};
 
 export default App;
